fix: rename entry to index.ts to match main
3.0.0 shipped no resolvable entry: main and types point at
dist/index.js but the entry was src/hyperaxe.ts.

Fixes #32
